move peer fetching to peer component
